+++ tests/db_sync_without_beautifulsoup_install_test.py
@@ -3,17 +3,14 @@ from typing import TYPE_CHECKING
 import pytest
 from pydantic import HttpUrl
 
-from yak_server.cli.database.sync import (
-    SyncOfficialResultsNotAvailableError,
-    synchronize_official_results,
-)
+from yak_server.cli.sync import SyncOfficialResultsNotAvailableError, synchronize_official_results
 
 if TYPE_CHECKING:
     from sqlalchemy import Engine
 
 
 def test_db_sync(monkeypatch: "pytest.MonkeyPatch", engine_for_test: "Engine") -> None:
-    monkeypatch.setattr("yak_server.cli.database.sync.bs4", None)
+    monkeypatch.setattr("yak_server.cli.sync.bs4", None)
 
     with pytest.raises(SyncOfficialResultsNotAvailableError) as exception:
         synchronize_official_results(engine_for_test, HttpUrl("https://fake.com"))
